pyproject.toml
[tool.poetry]
name = "shaped"
version = "0.1.0"
description = ""
authors = ["Maykon <maykon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
celery = "^5.2.7"
djangorestframework = "^3.13.1"
prettyconf = "^2.2.1"
gevent = "^21.12.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
